+++ feature/diary/src/main/java/com/boostcamp/dreamteam/dreamdiary/feature/diary/home/DiaryHomeScreen.kt
@@ -1,4 +1,4 @@
-package com.boostcamp.dreamteam.dreamdiary.feature.diary
+package com.boostcamp.dreamteam.dreamdiary.feature.diary.home
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
@@ -26,8 +26,8 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import com.boostcamp.dreamteam.dreamdiary.core.model.Diary
-import com.boostcamp.dreamteam.dreamdiary.feature.diary.components.DiaryListTab
-import com.boostcamp.dreamteam.dreamdiary.feature.diary.components.diariesPreview
+import com.boostcamp.dreamteam.dreamdiary.feature.diary.home.components.DiaryListTab
+import com.boostcamp.dreamteam.dreamdiary.feature.diary.home.components.diariesPreview
 
 @Composable
 fun DiaryHomeScreen() {
