+++ feature/diary/src/main/java/com/boostcamp/dreamteam/dreamdiary/feature/diary/home/components/DiaryCard.kt
@@ -1,4 +1,4 @@
-package com.boostcamp.dreamteam.dreamdiary.feature.diary.components
+package com.boostcamp.dreamteam.dreamdiary.feature.diary.home.components
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
