+++ feature/diary/src/main/java/com/boostcamp/dreamteam/dreamdiary/feature/diary/home/components/DiaryListTab.kt
@@ -1,4 +1,4 @@
-package com.boostcamp.dreamteam.dreamdiary.feature.diary.components
+package com.boostcamp.dreamteam.dreamdiary.feature.diary.home.components
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.PaddingValues
